requirements.txt
asgiref==3.5.2
awsebcli==3.20.10
boto3
botocore
cement==2.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.3
distlib==0.3.8
Django==4.1.2
django-admin-thumbnails==0.2.9
django-appconf==1.0.6
django-ipware==7.0.1
django-paypal==2.1
django-session-timeout==0.1.0
django-storages==1.14.4
filelock==3.15.4
gunicorn==23.0.0
idna==3.7
jmespath==1.0.1
packaging==24.1
pathspec==0.10.1
pillow==10.4.0
platformdirs==4.2.2
pypiwin32==223
python-dateutil==2.9.0.post0
python-decouple==3.4
python-ipware==3.0.0
pytz==2024.1
pywin32==306
PyYAML==6.0.2
requests==2.32.3
s3transfer
semantic-version==2.8.5
six==1.16.0
sqlparse==0.5.1
termcolor==1.1.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
wcwidth==0.1.9
